--- packages/visual-editor/src/internal/puck/constant-value-fields/ImageList.tsx
@@ -1,25 +1,0 @@
-import { pt } from "../../../utils/i18n/platform.ts";
-import { IMAGE_CONSTANT_CONFIG } from "./Image.tsx";
-import { getRandomPlaceholderImageObject } from "../../../utils/imagePlaceholders.ts";
-import { type YextArrayField } from "../../../editor/YextField.tsx";
-
-export const IMAGE_LIST_CONSTANT_CONFIG = (): YextArrayField<any> => {
-  return {
-    label: "",
-    type: "array",
-    arrayFields: {
-      assetImage: {
-        ...IMAGE_CONSTANT_CONFIG,
-        label: pt("fields.image", "Image"),
-      },
-    },
-    defaultItemProps: {
-      assetImage: {
-        ...getRandomPlaceholderImageObject({ width: 1000, height: 570 }),
-        width: 1000,
-        height: 570,
-      },
-    },
-    getItemSummary: (_, i) => pt("photo", "Photo") + " " + ((i ?? 0) + 1),
-  };
-};
